Add USB hub controller & dual stacked USB-C port
Now we can have a port to plug in our audio source, and an auxilliary
port for plugging in an additional device (such as a face tracker)!

The setup around USB2422 should ensure it initializes appropriately,
even when the RP2040 isn't set up to communicate with the USB2422,
with `*_DN1` being non-removable, and `*_DN2` being removable.

This will allow for firmware like sctanf's picoamp to be fully drop-in
compatible, and hopefully allow custom firmware to take advantage of
the USB2422's SMBus registers.