Replace NCP167 with NCP187
NCP187 lets us output 4.2V to our amplifiers with SNS feedback.
<https://www.onsemi.com/download/data-sheet/pdf/ncp187-d.pdf>

Per the amplifiers' datasheet, this increased VDD will reduce THD+N.
<https://www.analog.com/media/en/technical-documentation/data-sheets/max98360a-max98360d.pdf>

We choose 4.2V output because we can expect a minimum of 4.5V to be
delivered to the board per the USB 2.0 specification, and the NCP187
has a maximum dropout of 220mV.  4.5V - 220mV = 4.28V, so 4.2 it is.
<https://www.usb.org/document-library/usb-20-specification>

Setting the MAX98360 gains to 6dB should yield a peak voltage output of
3.67V, which is well within range of our 4.2V output.  This results in
up to 1.125W sent to each 6ohm speaker, which is within the 2W power
handling of the Tectonic TEBM35C02-6B's found on the Index, and does not
exceed USB-2.0's 2.5W supply power when both speakers are powered at
full volume.
<https://cdn.prod.website-files.com/673cf5977476dddd8360dc22/67779956ffce5474ffe874a4_TEBM35C02-6B%20Datasheet.pdf>

We'll just have to see if 1.125W is enough for adequate volume levels...